+++ authenticator/CHANGELOG.md
@@ -1,3 +1,26 @@
+## [Release 1.0.0](https://github.com/aws-amplify/amplify-ui-android/releases/tag/release_authenticator_v1.0.0)
+
+### Fixes
+- **authenticator:** Rename AuthenticatorMessage.Information to AuthenticatorMessage.Info ([#41](https://github.com/aws-amplify/amplify-ui-android/issues/41))
+
+[See all changes between 1.0.0-dev-preview.2 and 1.0.0](https://github.com/aws-amplify/amplify-ui-android/compare/release_authenticator_v1.0.0-dev-preview.2...release_authenticator_v1.0.0)
+
+## [Release 1.0.0-dev-preview.2](https://github.com/aws-amplify/amplify-ui-android/releases/tag/release_authenticator_v1.0.0-dev-preview.2)
+
+### Fixes
+- **authenticator:** Log auth exception ([#39](https://github.com/aws-amplify/amplify-ui-android/issues/39))
+
+[See all changes between 1.0.0-dev-preview.1 and 1.0.0-dev-preview.2](https://github.com/aws-amplify/amplify-ui-android/compare/release_authenticator_v1.0.0-dev-preview.1...release_authenticator_v1.0.0-dev-preview.2)
+
+## [Release 1.0.0-dev-preview.1](https://github.com/aws-amplify/amplify-ui-android/releases/tag/release_authenticator_v1.0.0-dev-preview.1)
+
+### Features
+- **authenticator:** Update dial code selector to a bottom sheet with search capability ([#32](https://github.com/aws-amplify/amplify-ui-android/issues/32))
+- **authenticator:** Add a material3 datepicker to date fields. ([#34](https://github.com/aws-amplify/amplify-ui-android/issues/34))
+- **authenticator:** Upgrade to Material 3 1.1.0 ([#31](https://github.com/aws-amplify/amplify-ui-android/issues/31))
+
+[See all changes between 1.0.0-dev-preview.0 and 1.0.0-dev-preview.1](https://github.com/aws-amplify/amplify-ui-android/compare/release_authenticator_v1.0.0-dev-preview.0...release_authenticator_v1.0.0-dev-preview.1)
+
 ## [Release 1.0.0-dev-preview.0](https://github.com/aws-amplify/amplify-ui-android/releases/tag/release_authenticator_v1.0.0-dev-preview.0)
 
 ### Features